% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-toolbox.R
\name{auto_recalculate}
\alias{auto_recalculate}
\title{Turn 'RAVE' module auto-recalculate on or off}
\usage{
auto_recalculate(on)
}
\arguments{
\item{on}{logical or integer; see details}
}
\description{
Turn 'RAVE' module auto-recalculate on or off
}
\details{
Used inside of 'RAVE' modules. By default modules are automatically
updated: that means if an input is updated, usually the module main function
will be executed and outputs will be rendered. However, if main body function
takes very long time to run, automatically recalculate module will result
in freeze shiny session. To solve the issue, module writers can choose
manually recalculate modules by turning off auto-recalculation.

\code{on} can be logical or numerical. For logical values, either the
auto-recalculation is constantly on or off. For numerical, typically
integer values, on means temporary on for limited times. For example
\code{auto_recalculate(1)} means run main function for one time.

To turn off auto-recalculation, call \code{auto_recalculate} within
\code{\link{define_initialization}} with \code{on=FALSE}. To temporary
turn on recalculation, use shiny \code{\link[shiny]{observeEvent}}
to capture an event and call \code{auto_recalculate(1)} once event is
triggered.
}
\examples{

require(shiny)
# Turn on debug mode for example
rave_context('rave_module_debug')

# In initialization
define_initialization({
  auto_recalculate(FALSE)
  # ..., other initialization code
})

}
