% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-afni_3dallineate.R
\name{cmd_run_3dAllineate}
\alias{cmd_run_3dAllineate}
\title{Align images using 'AFNI'}
\usage{
cmd_run_3dAllineate(
  subject,
  mri_path,
  ct_path,
  overwrite = FALSE,
  command_path = NULL,
  dry_run = FALSE,
  verbose = dry_run
)
}
\arguments{
\item{subject}{subject ID}

\item{ct_path, mri_path}{absolute paths to 'CT' and 'MR' image files}

\item{overwrite}{whether to overwrite existing files}

\item{command_path}{path to 'AFNI' home}

\item{dry_run}{dry-run flag}

\item{verbose}{whether to print out script}
}
\description{
This is a legacy script and possibly contain errors. Please use
\code{\link{cmd_run_ants_coreg}} for faster and stable implementation
instead.
}
