% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-compile_ui.R
\name{customizedUI}
\alias{customizedUI}
\title{Customized shiny widget for 'RAVE' modules}
\usage{
customizedUI(inputId, width = 12L, ...)
}
\arguments{
\item{inputId}{input ID used by \code{\link{define_input}},
or output ID, used by \code{\link{define_output}}}

\item{width}{width of the widget, used when it's output; see \code{width} in
\code{\link{define_output}}}

\item{...}{passed to \code{\link[shiny]{uiOutput}}}
}
\description{
Wrapper for shiny \code{\link[shiny]{uiOutput}}
}
