% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_compiler.R
\name{define_initialization}
\alias{define_initialization}
\title{Define module initialization expression}
\usage{
define_initialization(expr)
}
\arguments{
\item{expr}{R expression to run when initialize a module}
}
\description{
Used to define global variables and
change default behaviors of 'RAVE' modules. This function is a
\code{\link{rave_context}} generics and it's behavior will change when
running with/without shiny
}
