% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-compile_ui.R
\name{define_output}
\alias{define_output}
\title{Define 'RAVE' module outputs}
\usage{
define_output(
  definition,
  title = "",
  width = 12L,
  order = Inf,
  keyword = "outputId",
  watch_reactive = "input$..rave_output_update_all..",
  ...
)
}
\arguments{
\item{definition}{un-evaluated R expressions}

\item{title}{title of output}

\item{width}{integer from 1 to 12, width of output}

\item{order}{order of the output; output with smaller order will be at the
front row}

\item{keyword}{argument name for shiny output ID}

\item{watch_reactive}{reactive value to set once input is changed}

\item{...}{passed to other methods}
}
\description{
Define 'RAVE' module outputs
}
