% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-getters.R
\name{fake_session}
\alias{fake_session}
\title{Create a fake 'RAVE-shiny' session}
\usage{
fake_session(rave_id = "__fake_session__", id = NULL)
}
\arguments{
\item{rave_id}{internally used}

\item{id}{'RAVE' module ID}
}
\description{
Create a fake 'RAVE-shiny' session
}
