% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-volumetric_atlases.R
\name{generate_atlases_from_template}
\alias{generate_atlases_from_template}
\title{Create brain atlases from template}
\usage{
generate_atlases_from_template(
  subject,
  atlas_folders,
  template_name = "mni_icbm152_nlin_asym_09b",
  any_mni152 = TRUE,
  surfaces = TRUE,
  as_job = TRUE
)
}
\arguments{
\item{subject}{'RAVE' subject instance or character; see
\code{\link{as_rave_subject}}}

\item{atlas_folders}{paths to the atlas folders}

\item{template_name}{template name where the atlases are created;
see \code{\link{get_available_morph_to_template}} to get available templates
and \code{\link{cmd_run_yael_preprocess}} to generate transforms to the
templates}

\item{any_mni152}{if the template is 'MNI152', then whether to check other
template folders; default is true}

\item{surfaces}{whether to generate surfaces; if true, then a 'GIfTI' file
and a 'STL' file will be created for each atlas, if applicable. The
'GIfTI' file will be in right-anterior-superior 'RAS' coordinate system to
comply with the 'NIfTI' standard, and 'STL' will be in
left-posterior-superior 'LPS' system that can be imported to 'ITK'-based
software such as 'BrainLab' or 'ANTs'.}

\item{as_job}{whether to run as a background job for cleaner environment;
default is true; set to false for debugging.}
}
\value{
The function returns nothing, but will create a folder named
\code{'atlases'} under raw subject \code{'rave-imaging'} folder.
}
\description{
Reverse-transform from template atlases via volume mapping
}
\examples{

# Please check out https://rave.wiki to configure Python for RAVE
# or run ravemanager::configure_python()
\dontrun{

generate_atlases_from_template("YAEL/yael_demo_001")

}


}
