% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-volumetric_atlases.R
\name{get_available_morph_to_template}
\alias{get_available_morph_to_template}
\title{Get names of available non-linear transforms to the templates}
\usage{
get_available_morph_to_template(subject)
}
\arguments{
\item{subject}{'RAVE' subject instance or character; see
\code{\link{as_rave_subject}}}
}
\value{
Template names with valid non-linear transforms.
}
\description{
This function obtains existing transforms rather than creating new
transforms. Please check function \code{\link{cmd_run_yael_preprocess}}
about mapping native images to the templates. This function requires
additional 'Python' configuration.
}
\examples{

# Please check out https://rave.wiki to configure Python for RAVE
# or run ravemanager::configure_python()
\dontrun{

get_available_morph_to_template("project_name/subject_code")

}


}
