% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-getters.R
\name{getDefaultDataRepository}
\alias{getDefaultDataRepository}
\title{Get default data repository}
\usage{
getDefaultDataRepository()
}
\value{
environment
}
\description{
Default data repository is an environment shared by modules
and packages to store/fetch small global items. Not used often, but useful
when 'RAVE' modules need communication across packages.
}
