% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-getters.R
\name{getDefaultPackageData}
\alias{getDefaultPackageData}
\title{Get a list shared across modules within a same package}
\usage{
getDefaultPackageData()
}
\value{
A \code{\link[dipsaus]{fastmap2}} instance
}
\description{
Package data is a \code{\link[dipsaus]{fastmap2}} instance that
stores key-value pairs. A package data is shared across modules but
independent across 'RAVE' packages. It's useful to store shared data
for modules
}
