% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-getters.R
\name{getDefaultReactiveInput}
\alias{getDefaultReactiveInput}
\title{Get shiny \code{input} object}
\usage{
getDefaultReactiveInput(isolated = FALSE, ...)
}
\arguments{
\item{isolated}{whether to \code{\link[shiny]{isolate}} input as a list}

\item{...}{pass to other methods}
}
\description{
Get \code{session$input} object with correct scope
}
