% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-getters.R
\name{getDefaultReactiveOutput}
\alias{getDefaultReactiveOutput}
\title{Get shiny \code{output} object}
\usage{
getDefaultReactiveOutput(...)
}
\arguments{
\item{...}{pass to other methods}
}
\description{
Get \code{session$output} object with correct scope
}
