% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-voltage_repository.R
\name{glimpse-repository}
\alias{glimpse-repository}
\alias{glimpse_voltage_repository_with_blocks}
\title{Visualizes repositories with interactive plots}
\usage{
glimpse_voltage_repository_with_blocks(
  repository,
  initial_block = NULL,
  channels = NULL,
  epoch_name = NULL,
  start_time = 0,
  duration = 5,
  channel_gap = 1000,
  highpass_freq = NA,
  lowpass_freq = NA,
  shiny_ns = NULL
)
}
\arguments{
\item{repository}{'RAVE' repository}

\item{initial_block}{initial recording block to select}

\item{channels}{channels to visualize; default is all}

\item{epoch_name}{additional epoch to annotation}

\item{start_time, duration, channel_gap}{initial start time, duration, and
channel gap (can be changed later)}

\item{highpass_freq, lowpass_freq}{filter to apply when visualizing the
signals, useful when signals have 'DC' shift}

\item{shiny_ns}{shiny name-space}
}
\value{
An R-shiny application
}
\description{
Requires optional package \code{'plotly'}; please install the package prior
to launching the viewer.
}
\examples{

subject <- as_rave_subject("demo/DemoSubject", strict = FALSE)

if(interactive() && file.exists(subject$data_path)) {

  repository <- ravecore::prepare_subject_voltage_with_blocks(
    subject = subject
  )

  glimpse_voltage_repository_with_blocks(
    repository = repository,
    initial_block = "008",
    epoch_name = "auditory_onset",
    highpass_freq = 0.5
  )

}

}
