% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_compiler.R
\name{load_scripts}
\alias{load_scripts}
\title{Load external script in 'RAVE' modules}
\usage{
load_scripts(..., asis = TRUE)
}
\arguments{
\item{...}{file path or \code{rlang::quo}; see details.}

\item{asis}{for backward compatibility, ignored in current version.}
}
\description{
Load scripts such as output render functions or shiny observers
in \code{'comp.R'}
}
\details{
\code{'comp.R'} defines inputs and outputs, \code{'main.R'} defines
main function. For output render functions and customized reactive events,
store them in another file and load them via \code{load_scripts}. For
convenient purpose, short debug code don't have to sit in files: wrap
them in a \code{quosure}, and the expressions will be loaded during
compiling time.
}
