% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-paths.R
\name{module-package}
\alias{module-package}
\alias{rave_module_package}
\alias{rave_module_root_directory}
\alias{package_file}
\title{Get package information under current 'RAVE' context}
\usage{
rave_module_package()

rave_module_root_directory()

package_file(path, package = NULL)
}
\arguments{
\item{path}{relative path to file in the project}

\item{package}{package name}
}
\description{
Get package information under current 'RAVE' context
}
\examples{

# Debug mode, but package is not specified
rave_context('rave_module_debug')
rave_module_package()

# Debug under ravecore package
rave_context('rave_module_debug', package = 'ravecore')
rave_module_package()

# Production mode, running modules under 'ravecore'
# always use with_rave_context to temporary set rave_running context
with_rave_context(context = 'rave_running', {
  rave_module_package()
}, package = 'ravecore')


# Get package directory
rave_context('rave_module_debug', package = 'raveutils')
rave_module_root_directory()

# Get path to a file within the package
# When debugging ravecore, returns relative path of input_compound.js
# When running in production mode, returns system path
package_file('./inst/assets/input_compound.js', package = 'ravecore')

}
\seealso{
\code{\link{rave-context}}
}
