% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-shiny_observer.R
\name{module_notification}
\alias{module_notification}
\title{Safe way to show shiny notifications}
\usage{
module_notification(
  ...,
  type = c("message", "warning", "error", "default"),
  duration = 10,
  closeBotton = TRUE,
  action = NULL,
  id,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{messages to display}

\item{type}{message type, choices are "message", "warning", "error",
"default"}

\item{duration}{seconds for notification to stay}

\item{closeBotton, action, id, session}{passed to
\code{\link[shiny]{showNotification}}}
}
\description{
Show notification when shiny is running, and show console
messages when shiny is offline
}
