% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-shiny_observer.R
\name{module_remove_notification}
\alias{module_remove_notification}
\title{Safe way to remove shiny notification within module}
\usage{
module_remove_notification(id, ...)
}
\arguments{
\item{id, ...}{passed to \code{\link[shiny]{removeNotification}}}
}
\description{
Safe way to remove shiny notification within module
}
