% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-tools.R
\name{rave_cmd_tools}
\alias{rave_cmd_tools}
\alias{normalize_commandline_path}
\alias{cmd_dcm2niix}
\alias{cmd_freesurfer_home}
\alias{cmd_fsl_home}
\alias{cmd_afni_home}
\alias{cmd_homebrew}
\alias{cmd_dry_run}
\alias{rscript_path}
\title{Find external command-line tools}
\usage{
normalize_commandline_path(
  path,
  type = c("dcm2niix", "freesurfer", "fsl", "afni", "others"),
  unset = NA
)

cmd_dcm2niix(error_on_missing = TRUE, unset = NA)

cmd_freesurfer_home(error_on_missing = TRUE, unset = NA)

cmd_fsl_home(error_on_missing = TRUE, unset = NA)

cmd_afni_home(error_on_missing = TRUE, unset = NA)

cmd_homebrew(error_on_missing = TRUE, unset = NA)

cmd_dry_run()

rscript_path(...)
}
\arguments{
\item{path}{path to normalize}

\item{type}{type of command}

\item{unset}{default to return if the command is not found}

\item{error_on_missing}{whether to raise errors if command is missing}

\item{...}{ignored}
}
\value{
Normalized path to the command, or \code{unset} if command is
missing.
}
\description{
Find external command-line tools
}
