% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class001-base.R
\name{rave_path}
\alias{rave_path}
\title{Find file paths based on storage}
\usage{
rave_path(x, storage = NULL, ...)
}
\arguments{
\item{x}{R object}

\item{storage}{storage type, different options based on different R objects}

\item{...}{additional arguments passed to dispatched method}
}
\description{
A generic function that will be dispatched to using different method based
on input \code{x}
}
