% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-constants.R
\name{ravecore-constants}
\alias{ravecore-constants}
\alias{LOCATION_TYPES}
\alias{SIGNAL_TYPES}
\alias{IMPORT_FORMATS}
\alias{YAEL_IMAGE_TYPES}
\alias{MNI305_to_MNI152}
\title{'RAVE' constants}
\usage{
LOCATION_TYPES

SIGNAL_TYPES

IMPORT_FORMATS

YAEL_IMAGE_TYPES

MNI305_to_MNI152
}
\description{
Constant variables
}
