% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravecore-package.R
\docType{package}
\name{ravecore-package}
\alias{ravecore}
\alias{ravecore-package}
\title{ravecore: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Zhengjia Wang \email{dipterix.wang@gmail.com}

}
\keyword{internal}
