% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-module.R
\name{RAVEPackage}
\alias{RAVEPackage}
\title{Definition for 'RAVE' package instance}
\description{
Definition for 'RAVE' package instance

Definition for 'RAVE' package instance
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conf_path}}{configuration path}

\item{\code{package_name}}{R package name}

\item{\code{package_data}}{package data; see also \code{\link{getDefaultPackageData}}}

\item{\code{package_env}}{package name space}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RAVEPackage$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEPackage$new(package)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{package}}{'RAVE' package name}
}
\if{html}{\out{</div>}}
}
}
}
