% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-repository.R
\name{RAVERepository}
\alias{RAVERepository}
\title{Definition for data repository class}
\description{
Do not create instance directly, use
\code{\link{loaded_rave_repository}} instead.
}
\seealso{
\code{\link{rave_brain2}}, \code{\link[threeBrain]{freesurfer_brain2}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subject}}{\code{\link{RAVESubject}} instance}

\item{\code{electrodes}}{electrodes loaded}

\item{\code{ignored_electrodes}}{electrodes ignored}

\item{\code{reference_table}}{reference table}

\item{\code{time_range}}{seconds before and after trial onset}

\item{\code{reference_name}}{name of reference table}

\item{\code{epoch}}{\code{\link{RAVEEpoch}} instance}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{signature}}{unique signature for repository; if subject ID,
reference table, epoch table, and time range are the same, signatures
will be the same}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RAVERepository$new()}}
\item \href{#method-load_brain}{\code{RAVERepository$load_brain()}}
\item \href{#method-epoch_power}{\code{RAVERepository$epoch_power()}}
\item \href{#method-clear_cache}{\code{RAVERepository$clear_cache()}}
\item \href{#method-clear_memory}{\code{RAVERepository$clear_memory()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVERepository$new(subject, reference, epoch, before_onset, after_onset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject}}{character or \code{\link{RAVESubject}} instance}

\item{\code{reference}}{reference name}

\item{\code{epoch}}{epoch name}

\item{\code{before_onset}}{seconds to be loaded before trial onset}

\item{\code{after_onset}}{seconds to be loaded after trial onset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_brain"></a>}}
\if{latex}{\out{\hypertarget{method-load_brain}{}}}
\subsection{Method \code{load_brain()}}{
load 3D brain instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVERepository$load_brain(
  surfaces = "pial",
  use_141 = TRUE,
  recache = FALSE,
  clean_before_cache = FALSE,
  compute_template = FALSE,
  usetemplateifmissing = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surfaces}}{surface types such as \code{'pial'}, \code{'white'},
\code{'smoothwm'}, \code{'sphere'}, etc.}

\item{\code{use_141}}{whether to use 'SUMA' 141 standard brain if possible}

\item{\code{recache}}{whether to force re-calculate cache data}

\item{\code{clean_before_cache}}{whether to clear data before caching}

\item{\code{compute_template}}{whether to compute template vertices}

\item{\code{usetemplateifmissing}}{whether to use template brain if surfaces
are missing, default is false}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{threeBrain} brain instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epoch_power"></a>}}
\if{latex}{\out{\hypertarget{method-epoch_power}{}}}
\subsection{Method \code{epoch_power()}}{
epoch power data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVERepository$epoch_power(electrodes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{electrodes}}{electrodes to epoch}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
call \code{clear_cache} for each electrodes loaded
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVERepository$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_memory"></a>}}
\if{latex}{\out{\hypertarget{method-clear_memory}{}}}
\subsection{Method \code{clear_memory()}}{
call \code{clear_memory} for each electrodes loaded
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVERepository$clear_memory()}\if{html}{\out{</div>}}
}

}
}
