% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class005-subject.R
\name{RAVESubject}
\alias{RAVESubject}
\title{Defines 'RAVE' subject class}
\description{
\code{R6} class definition
}
\seealso{
\code{\link{load_meta2}}
}
\section{Super class}{
\code{\link[ravepipeline:RAVESerializable]{ravepipeline::RAVESerializable}} -> \code{RAVESubject}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{@impl}}{the internal object}

    \item{\code{project}}{project instance of current subject; see
\code{\link{RAVEProject}}}

    \item{\code{project_name}}{character string of project name}

    \item{\code{subject_code}}{character string of subject code}

    \item{\code{subject_id}}{subject ID: \code{"project/subject"}}

    \item{\code{path}}{subject root path}

    \item{\code{rave_path}}{'rave' directory under subject root path}

    \item{\code{meta_path}}{meta data directory for current subject}

    \item{\code{imaging_path}}{root path to imaging processing folder}

    \item{\code{freesurfer_path}}{'FreeSurfer' directory for current subject. If
no path exists, values will be \code{NA}}

    \item{\code{preprocess_path}}{preprocess directory under subject 'rave' path}

    \item{\code{data_path}}{data directory under subject 'rave' path}

    \item{\code{cache_path}}{path to 'FST' copies under subject 'data' path}

    \item{\code{pipeline_path}}{path to pipeline scripts under subject's folder}

    \item{\code{report_path}}{path to pipeline scripts under subject's folder}

    \item{\code{note_path}}{path that stores 'RAVE' related subject notes}

    \item{\code{epoch_names}}{possible epoch names}

    \item{\code{reference_names}}{possible reference names}

    \item{\code{reference_path}}{reference path under 'rave' folder}

    \item{\code{preprocess_settings}}{preprocess instance; see
\code{\link{RAVEPreprocessSettings}}}

    \item{\code{blocks}}{subject experiment blocks in current project}

    \item{\code{electrodes}}{all electrodes, no matter excluded or not}

    \item{\code{raw_sample_rates}}{voltage sample rate}

    \item{\code{power_sample_rate}}{power spectrum sample rate}

    \item{\code{has_wavelet}}{whether electrodes have wavelet transforms}

    \item{\code{notch_filtered}}{whether electrodes are Notch-filtered}

    \item{\code{electrode_types}}{electrode signal types}

    \item{\code{electrode_composed}}{composed electrode channels, not actual
physically contacts, but is generated from those physically ones}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RAVESubject-@marshal}{\code{RAVESubject$@marshal()}}
    \item \href{#method-RAVESubject-@unmarshal}{\code{RAVESubject$@unmarshal()}}
    \item \href{#method-RAVESubject-print}{\code{RAVESubject$print()}}
    \item \href{#method-RAVESubject-format}{\code{RAVESubject$format()}}
    \item \href{#method-RAVESubject-initialize}{\code{RAVESubject$new()}}
    \item \href{#method-RAVESubject-meta_data}{\code{RAVESubject$meta_data()}}
    \item \href{#method-RAVESubject-valid_electrodes}{\code{RAVESubject$valid_electrodes()}}
    \item \href{#method-RAVESubject-initialize_paths}{\code{RAVESubject$initialize_paths()}}
    \item \href{#method-RAVESubject-set_default}{\code{RAVESubject$set_default()}}
    \item \href{#method-RAVESubject-get_default}{\code{RAVESubject$get_default()}}
    \item \href{#method-RAVESubject-get_note_summary}{\code{RAVESubject$get_note_summary()}}
    \item \href{#method-RAVESubject-get_epoch}{\code{RAVESubject$get_epoch()}}
    \item \href{#method-RAVESubject-get_reference}{\code{RAVESubject$get_reference()}}
    \item \href{#method-RAVESubject-get_electrode_table}{\code{RAVESubject$get_electrode_table()}}
    \item \href{#method-RAVESubject-get_frequency}{\code{RAVESubject$get_frequency()}}
    \item \href{#method-RAVESubject-list_pipelines}{\code{RAVESubject$list_pipelines()}}
    \item \href{#method-RAVESubject-load_pipeline}{\code{RAVESubject$load_pipeline()}}
    \item \href{#method-RAVESubject-clone}{\code{RAVESubject$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ravepipeline" data-topic="RAVESerializable" data-id="@compare"><a href='../../ravepipeline/html/RAVESerializable.html#method-RAVESerializable-@compare'><code>ravepipeline::RAVESerializable$@compare()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-@marshal"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-@marshal}{}}}
\subsection{\code{RAVESubject$@marshal()}}{
  Internal method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$@marshal(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{internal arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-@unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-@unmarshal}{}}}
\subsection{\code{RAVESubject$@unmarshal()}}{
  Internal method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$@unmarshal(object, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{object, ...}}{internal arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-print"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-print}{}}}
\subsection{\code{RAVESubject$print()}}{
  override print method: prints \code{format()}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{passed to \code{format()}}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-format"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-format}{}}}
\subsection{\code{RAVESubject$format()}}{
  summary of the subject's data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$format(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{ignored}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A character vector, one element per line
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-initialize}{}}}
\subsection{\code{RAVESubject$new()}}{
  constructor
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$new(
  project_name,
  subject_code = NULL,
  strict = TRUE,
  parent_path = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{project_name}}{character project name}
      \item{\code{subject_code}}{character subject code}
      \item{\code{strict}}{whether to check if subject folders exist}
      \item{\code{parent_path}}{parent path if no default path is used, this is for
the root directory if subject is in 'BIDS' format}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-meta_data"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-meta_data}{}}}
\subsection{\code{RAVESubject$meta_data()}}{
  get subject meta data located in \code{"meta/"} folder
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$meta_data(
  meta_type = c("electrodes", "frequencies", "time_points", "epoch", "references"),
  meta_name = "default",
  strict = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{meta_type}}{choices are 'electrodes', 'frequencies', 'time_points',
'epoch', 'references'}
      \item{\code{meta_name}}{if \code{meta_type='epoch'}, read in
\code{'epoch_<meta_name>.csv'}; if \code{meta_type='references'},
read in \code{'reference_<meta_name>.csv'}.}
      \item{\code{strict}}{whether to raise errors if the files are missing; default
is true; alternative is to return \code{NULL} on missing}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    data frame
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-valid_electrodes"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-valid_electrodes}{}}}
\subsection{\code{RAVESubject$valid_electrodes()}}{
  get valid electrode numbers
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$valid_electrodes(reference_name = NULL, refresh = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{reference_name}}{character, reference name, see \code{meta_name}
in \code{self$meta_data} or \code{\link{load_meta2}} when
\code{meta_type} is 'reference'}
      \item{\code{refresh}}{whether to reload reference table before obtaining data,
default is false}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    integer vector of valid electrodes
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-initialize_paths"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-initialize_paths}{}}}
\subsection{\code{RAVESubject$initialize_paths()}}{
  create subject's directories on hard disk
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$initialize_paths(include_freesurfer = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{include_freesurfer}}{whether to create 'FreeSurfer' path}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-set_default"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-set_default}{}}}
\subsection{\code{RAVESubject$set_default()}}{
  set default key-value pair for the subject, used by 'RAVE'
modules
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$set_default(key, value, namespace = "default")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{character}
      \item{\code{value}}{value of the key}
      \item{\code{namespace}}{file name of the note (without post-fix)}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The same as \code{value}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-get_default"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-get_default}{}}}
\subsection{\code{RAVESubject$get_default()}}{
  get default key-value pairs for the subject, used by 'RAVE'
modules
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$get_default(
  ...,
  default_if_missing = NULL,
  simplify = TRUE,
  namespace = "default"
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{single key, or a vector of character keys}
      \item{\code{default_if_missing}}{default value is any key is missing}
      \item{\code{simplify}}{whether to simplify the results if there is only one key
to fetch; default is \code{TRUE}}
      \item{\code{namespace}}{file name of the note (without post-fix)}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A named list of key-value pairs, or if one key is specified and
\code{simplify=TRUE}, then only the value will be returned.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-get_note_summary"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-get_note_summary}{}}}
\subsection{\code{RAVESubject$get_note_summary()}}{
  get summary table of all the key-value pairs used by 'RAVE'
modules for the subject
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$get_note_summary(namespaces, include_history = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{namespaces}}{namespaces for the entries; see method
\code{get_default} or \code{set_default}. Default is all possible
namespaces}
      \item{\code{include_history}}{whether to include history entries; default is
false}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A data frame with four columns: \code{'namespace'} for the group
name of the entry (entries within the same namespace usually share same
module), \code{'timestamp'} for when the entry was registered.
\code{'entry_name'} is the name of the entry. If \code{include_history}
is true, then multiple entries with the same \code{'entry_name'} might
appear since the obsolete entries are included. \code{'entry_value'}
is the value of the corresponding entry.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-get_epoch"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-get_epoch}{}}}
\subsection{\code{RAVESubject$get_epoch()}}{
  check and get subject's epoch information
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$get_epoch(
  epoch_name = "default",
  as_table = FALSE,
  trial_starts = 0
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{epoch_name}}{epoch name, depending on the subject's meta files}
      \item{\code{as_table}}{whether to convert to \code{\link{data.frame}}; default
is false}
      \item{\code{trial_starts}}{the start of the trial relative to epoch time;
default is 0}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    If \code{as_table} is \code{FALSE}, then returns as
\code{\link{RAVEEpoch}} instance; otherwise returns epoch table; will
raise errors when file is missing or the epoch is invalid.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-get_reference"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-get_reference}{}}}
\subsection{\code{RAVESubject$get_reference()}}{
  check and get subject's reference information
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$get_reference(reference_name, simplify = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{reference_name}}{reference name, depending on the subject's meta
file settings}
      \item{\code{simplify}}{whether to only return the reference column}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    If \code{simplify} is true, returns a vector of reference
electrode names, otherwise returns the whole table; will
raise errors when file is missing or the reference is invalid.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-get_electrode_table"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-get_electrode_table}{}}}
\subsection{\code{RAVESubject$get_electrode_table()}}{
  check and get subject's electrode table with electrodes
that are load-able
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$get_electrode_table(
  electrodes,
  reference_name,
  subset = FALSE,
  simplify = FALSE,
  warn = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{electrodes}}{characters indicating integers such as
\code{"1-14,20-30"}, or integer vector of electrode numbers}
      \item{\code{reference_name}}{see method \code{get_reference}}
      \item{\code{subset}}{whether to subset the resulting data table}
      \item{\code{simplify}}{whether to only return electrodes}
      \item{\code{warn}}{whether to warn about missing electrodes; default is true}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    If \code{simplify} is true, returns a vector of electrodes
that are valid (or won't be excluded) under given reference; otherwise
returns a table. If \code{subset} is true, then the table will be
subset and only rows with electrodes to be loaded will be kept.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-get_frequency"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-get_frequency}{}}}
\subsection{\code{RAVESubject$get_frequency()}}{
  check and get subject's frequency table, time-frequency
decomposition is needed.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$get_frequency(simplify = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{simplify}}{whether to simplify as vector}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    If \code{simplify} is true, returns a vector of frequencies;
otherwise returns a table.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-list_pipelines"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-list_pipelines}{}}}
\subsection{\code{RAVESubject$list_pipelines()}}{
  list saved pipelines
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$list_pipelines(
  pipeline_name,
  cache = FALSE,
  check = TRUE,
  all = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pipeline_name}}{pipeline ID}
      \item{\code{cache}}{whether to use cache registry to speed up}
      \item{\code{check}}{whether to check if the pipelines exist}
      \item{\code{all}}{whether to list all pipelines; default is false; pipelines
with the same label but older time-stamps will be hidden}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A table of pipeline registry
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-load_pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-load_pipeline}{}}}
\subsection{\code{RAVESubject$load_pipeline()}}{
  load saved pipeline
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$load_pipeline(directory)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{directory}}{pipeline directory name}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{PipelineTools} instance
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubject-clone}{}}}
\subsection{\code{RAVESubject$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubject$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
