% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class100-base_repository.R
\name{RAVESubjectBaseRepository}
\alias{RAVESubjectBaseRepository}
\title{'RAVE' class for base repository}
\description{
The class is for creating child classes, to instantiate the class,
please use \code{\link{prepare_subject_bare0}} to create base repository.
}
\seealso{
\code{\link{prepare_subject_bare0}}
}
\section{Super class}{
\code{\link[ravepipeline:RAVESerializable]{ravepipeline::RAVESerializable}} -> \code{RAVESubjectRepository}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{@restored}}{internal flag indicating whether the repository
is restored from serialization. Repositories restored from serialization
will behave differently (slightly) for performance considerations}

    \item{\code{repository_id}}{repository identifier, typically generated with
random string}

    \item{\code{quiet}}{whether to suppress update warning messages, when
requested electrodes are not fully processed or excluded}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{auto_exclude}}{whether to automatically discard channels that are
marked as "excluded" (such as bad channels or channels that should
not be analyzed); default is often true}

    \item{\code{meta_info}}{list of meta information}

    \item{\code{needs_update}}{write-only attribute when subject needs to be
reloaded from the disk and reference table needs to be updated, use
\code{repo$needs_update <- TRUE}}

    \item{\code{project}}{project instance, see \code{\link{RAVEProject}}}

    \item{\code{subject}}{subject instance, see \code{\link{RAVESubject}}}

    \item{\code{electrode_list}}{integer vector of electrodes included}

    \item{\code{electrode_table}}{the entire electrode table}

    \item{\code{electrode_signal_types}}{more accurate name should be "channel"
signal types: currently returns \code{'LFP'}, \code{'Auxiliary'}, or
\code{'Spike'}, for each channel}

    \item{\code{electrode_instances}}{electrode channel instance helpers for
loading electrode data}

    \item{\code{reference_name}}{name of reference table}

    \item{\code{reference_table}}{reference table}

    \item{\code{references_list}}{a vector of reference channel names, used together
with \code{reference_instances}}

    \item{\code{reference_instances}}{instances of reference channels, for
referencing on the fly, used for \code{electrode_instances}}

    \item{\code{digest_key}}{a list of repository data used to generate
repository signature}

    \item{\code{signature}}{signature of the repository, two repositories might
share the same signature if their contents are the same (even with
different identifiers); generated from \code{digest_key}}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RAVESubjectRepository-@get_container}{\code{RAVESubjectRepository$@get_container()}}
    \item \href{#method-RAVESubjectRepository-@marshal}{\code{RAVESubjectRepository$@marshal()}}
    \item \href{#method-RAVESubjectRepository-@unmarshal}{\code{RAVESubjectRepository$@unmarshal()}}
    \item \href{#method-RAVESubjectRepository-print}{\code{RAVESubjectRepository$print()}}
    \item \href{#method-RAVESubjectRepository-initialize}{\code{RAVESubjectRepository$new()}}
    \item \href{#method-RAVESubjectRepository-export_matlab}{\code{RAVESubjectRepository$export_matlab()}}
    \item \href{#method-RAVESubjectRepository-get_electrode_coordinate}{\code{RAVESubjectRepository$get_electrode_coordinate()}}
    \item \href{#method-RAVESubjectRepository-clone}{\code{RAVESubjectRepository$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ravepipeline" data-topic="RAVESerializable" data-id="@compare"><a href='../../ravepipeline/html/RAVESerializable.html#method-RAVESerializable-@compare'><code>ravepipeline::RAVESerializable$@compare()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-@get_container"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-@get_container}{}}}
\subsection{\code{RAVESubjectRepository$@get_container()}}{
  Internal method, do not use it directly
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$@get_container()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-@marshal"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-@marshal}{}}}
\subsection{\code{RAVESubjectRepository$@marshal()}}{
  Internal method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$@marshal(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{internal arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-@unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-@unmarshal}{}}}
\subsection{\code{RAVESubjectRepository$@unmarshal()}}{
  Internal method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$@unmarshal(object, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{object, ...}}{internal arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-print"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-print}{}}}
\subsection{\code{RAVESubjectRepository$print()}}{
  User-friendly print method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$print()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-initialize}{}}}
\subsection{\code{RAVESubjectRepository$new()}}{
  constructor
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$new(
  subject,
  electrodes = NULL,
  reference_name = NULL,
  ...,
  auto_exclude = TRUE,
  quiet = TRUE,
  repository_id = NULL,
  strict = TRUE,
  .class = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{subject}}{'RAVE' subject}
      \item{\code{electrodes}}{string or integers indicating electrodes to load}
      \item{\code{reference_name}}{name of the reference table}
      \item{\code{...}}{reserved, currently ignored}
      \item{\code{auto_exclude}}{whether to automatically discard bad channels}
      \item{\code{quiet}}{see field \code{quiet}}
      \item{\code{repository_id}}{see field \code{repository_id}}
      \item{\code{strict}}{whether the mode should be strict; default is true and
errors out when subject is missing}
      \item{\code{.class}}{internally used, do not set, even if you know what this is}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-export_matlab"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-export_matlab}{}}}
\subsection{\code{RAVESubjectRepository$export_matlab()}}{
  Export the repository to 'Matlab' for future analysis
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$export_matlab(..., verbose = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{reserved for child classes}
      \item{\code{verbose}}{print progresses}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The root directory where the files are stored.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-get_electrode_coordinate"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-get_electrode_coordinate}{}}}
\subsection{\code{RAVESubjectRepository$get_electrode_coordinate()}}{
  Get loaded electrode coordinates by type
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$get_electrode_coordinate(
  type = "LFP",
  electrodes = KEY_MISSING,
  strict = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{type}}{length of one or more, electrode channel (signal) type;
see constant \code{SIGNAL_TYPES}. Default is \code{'LFP'}}
      \item{\code{electrodes}}{electrodes to subset, default is a missing key, which
uses all the loaded electrodes to subset}
      \item{\code{strict}}{whether the returned table must not be empty; default is
true, which will raise errors when no electrode matches}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Filtered table; see \code{\link{load_meta2}} with meta type
\code{'electrodes'}. Besides the mandatory columns, the returned table
also contains \code{'LabelPrefix'}, used by \pkg{YAEL} to store surgical
labels to group channels by electrode device; \code{'LeadChannel'}
(logical) to indicate whether the channel is a leading channel in this
group, typically the inner-most channel; and \code{'ShortLabel'} for
short electrode label.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRepository-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRepository-clone}{}}}
\subsection{\code{RAVESubjectRepository$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRepository$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
