% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class115-recording_blocks_power.R
\name{RAVESubjectRecordingBlockPowerRepository}
\alias{RAVESubjectRecordingBlockPowerRepository}
\title{'RAVE' class for loading time-frequency power components}
\description{
Loads time-frequency power
}
\seealso{
\code{\link{prepare_subject_power_with_blocks}}
}
\section{Super classes}{
\code{\link[ravepipeline:RAVESerializable]{ravepipeline::RAVESerializable}} -> \code{RAVESubjectRepository} -> \code{\link[ravecore:RAVESubjectRecordingBlockRepository]{RAVESubjectRecordingBlockRepository}} -> \code{\link[ravecore:RAVESubjectRecordingBlockTimeFreqBaseRepository]{RAVESubjectRecordingBlockTimeFreqBaseRepository}} -> \code{RAVESubjectRecordingBlockPowerRepository}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{power}}{data container, alias of \code{get_container}}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RAVESubjectRecordingBlockPowerRepository-@marshal}{\code{RAVESubjectRecordingBlockPowerRepository$@marshal()}}
    \item \href{#method-RAVESubjectRecordingBlockPowerRepository-@unmarshal}{\code{RAVESubjectRecordingBlockPowerRepository$@unmarshal()}}
    \item \href{#method-RAVESubjectRecordingBlockPowerRepository-initialize}{\code{RAVESubjectRecordingBlockPowerRepository$new()}}
    \item \href{#method-RAVESubjectRecordingBlockPowerRepository-clone}{\code{RAVESubjectRecordingBlockPowerRepository$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ravepipeline" data-topic="RAVESerializable" data-id="@compare"><a href='../../ravepipeline/html/RAVESerializable.html#method-RAVESerializable-@compare'><code>ravepipeline::RAVESerializable$@compare()</code></a></span></li>
  <li><code>RAVESubjectRepository$@get_container()</code></li>
  <li><code>RAVESubjectRepository$get_electrode_coordinate()</code></li>
  <li><code>RAVESubjectRepository$print()</code></li>
  <li><span class="pkg-link" data-pkg="ravecore" data-topic="RAVESubjectRecordingBlockRepository" data-id="export_matlab"><a href='../../ravecore/html/RAVESubjectRecordingBlockRepository.html#method-RAVESubjectRecordingBlockRepository-export_matlab'><code>RAVESubjectRecordingBlockRepository$export_matlab()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ravecore" data-topic="RAVESubjectRecordingBlockRepository" data-id="get_container"><a href='../../ravecore/html/RAVESubjectRecordingBlockRepository.html#method-RAVESubjectRecordingBlockRepository-get_container'><code>RAVESubjectRecordingBlockRepository$get_container()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ravecore" data-topic="RAVESubjectRecordingBlockTimeFreqBaseRepository" data-id="mount_data"><a href='../../ravecore/html/RAVESubjectRecordingBlockTimeFreqBaseRepository.html#method-RAVESubjectRecordingBlockTimeFreqBaseRepository-mount_data'><code>RAVESubjectRecordingBlockTimeFreqBaseRepository$mount_data()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRecordingBlockPowerRepository-@marshal"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRecordingBlockPowerRepository-@marshal}{}}}
\subsection{\code{RAVESubjectRecordingBlockPowerRepository$@marshal()}}{
  Internal method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRecordingBlockPowerRepository$@marshal(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{internal arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRecordingBlockPowerRepository-@unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRecordingBlockPowerRepository-@unmarshal}{}}}
\subsection{\code{RAVESubjectRecordingBlockPowerRepository$@unmarshal()}}{
  Internal method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRecordingBlockPowerRepository$@unmarshal(object, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{object, ...}}{internal arguments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRecordingBlockPowerRepository-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRecordingBlockPowerRepository-initialize}{}}}
\subsection{\code{RAVESubjectRecordingBlockPowerRepository$new()}}{
  constructor
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRecordingBlockPowerRepository$new(
  subject,
  electrodes = NULL,
  reference_name = NULL,
  blocks = NULL,
  ...,
  quiet = FALSE,
  repository_id = NULL,
  strict = TRUE,
  lazy_load = FALSE,
  .class = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{subject}}{'RAVE' subject}
      \item{\code{electrodes}}{string or integers indicating electrodes to load}
      \item{\code{reference_name}}{name of the reference table}
      \item{\code{blocks}}{name of the recording blocks to load}
      \item{\code{...}}{passed to \code{\link{RAVESubjectBaseRepository}} constructor}
      \item{\code{quiet}}{see field \code{quiet}}
      \item{\code{repository_id}}{see field \code{repository_id}}
      \item{\code{strict}}{whether the mode should be strict; default is true and
errors out when subject is missing}
      \item{\code{lazy_load}}{whether to delay (lazy) the evaluation \code{mount_data}}
      \item{\code{.class}}{internally used, do not set, even if you know what this is}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RAVESubjectRecordingBlockPowerRepository-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RAVESubjectRecordingBlockPowerRepository-clone}{}}}
\subsection{\code{RAVESubjectRecordingBlockPowerRepository$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RAVESubjectRecordingBlockPowerRepository$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
