% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class003-subject-internal.R
\name{save_meta2}
\alias{save_meta2}
\title{Function to save meta data to 'RAVE' subject}
\usage{
save_meta2(data, meta_type, project_name, subject_code)
}
\arguments{
\item{data}{data table}

\item{meta_type}{see load meta}

\item{project_name}{project name}

\item{subject_code}{subject code}
}
\value{
Either none if no meta matched or the absolute path of file saved.
}
\description{
Function to save meta data to 'RAVE' subject
}
\examples{

subject <- as_rave_subject("demo/DemoSubject", strict = FALSE)

if(interactive() && dir.exists(subject$meta_path)) {

  electrode_table <- subject$get_electrode_table()

  save_meta2(
    "electrodes",
    electrode_table,
    project_name = subject$project_name,
    subject_code = subject$subject_code
  )
}

}
