% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-snapshot.R
\name{snapshot_project}
\alias{snapshot_project}
\alias{snapshot_subject}
\title{Create overview report for given project or subject}
\usage{
snapshot_subject(x, target_path = NULL, quick = FALSE)

snapshot_project(
  x,
  target_path = NULL,
  template_subjects = NULL,
  quick = FALSE
)
}
\arguments{
\item{x}{characters or a 'RAVE' project or subject instance}

\item{target_path}{directory where the snapshot data will be stored; default
is under \code{'project_overview'} group data folder, see method
\code{group_data} from \code{\link{RAVEProject}} class.}

\item{quick}{whether to skip certain validations and subjects if snapshot
reports have already existed}

\item{template_subjects}{a vector of characters of template brain to be
used for generating the group brain; see
\code{\link[threeBrain]{available_templates}} for available templates.}
}
\value{
\code{snapshot_subject} returns a list of subject summary and
the calculated target path; \code{snapshot_project} returns the path to
the compiled project report in 'HTML'.
}
\description{
Create overview report for given project or subject
}
\examples{

\dontrun{

# Run in parallel
ravepipeline::with_rave_parallel({

  snapshot_project("demo")

})

}

}
