% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-modules.R, R/core-single_app.R
\name{start-app}
\alias{start-app}
\alias{start_rave}
\alias{view_layout}
\title{Start 'RAVE' Applications}
\usage{
start_rave(
  host = "127.0.0.1",
  port = NULL,
  launch_browser = TRUE,
  test_mode = FALSE,
  token = NULL,
  theme = "purple",
  .adapter = dipsaus::fastmap2()
)

view_layout(host = "127.0.0.1", port = NULL, theme = "red")
}
\arguments{
\item{host}{IP address, default is \code{'127.0.0.1'}}

\item{port}{integer port to listen to, default is \code{NULL}, and will be
assigned a random port}

\item{launch_browser}{whether to launch browser once the application is
opened}

\item{test_mode}{whether to enter test mode. Test mode does not release
memory once application is closed, this will allow step-by-step debugging}

\item{token}{token needed to launch application, default does not require
any token}

\item{theme}{passed to \code{\link[shinydashboard]{dashboardPage}}}

\item{.adapter}{debug use, must be a \code{\link[dipsaus]{fastmap2}}
instance. Combined with \code{test_mode=TRUE}, \code{.adapter} stores
all container/module instances.}
}
\description{
Use \code{start_rave} to launch applications; use
\code{view_layout} to debug module.
}
\examples{

if(interactive()){

  # Production
  start_rave()

  # Debug rave preprocess
  rave_context('rave_module_debug', package = 'ravetools',
               module_id = 'ravepreprocessoverview')
  view_layout()
}

}
