% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-subject.R
\name{as_rave_project}
\alias{as_rave_project}
\title{Convert character to \code{\link{RAVEProject}} instance}
\usage{
as_rave_project(project)
}
\arguments{
\item{project}{character project name}
}
\description{
Convert character to \code{\link{RAVEProject}} instance
}
\seealso{
\code{\link{RAVEProject}}
}
