% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-volumetric_atlases.R
\name{generate_atlases_from_template}
\alias{generate_atlases_from_template}
\title{Create brain atlases from template}
\usage{
generate_atlases_from_template(
  subject,
  atlas_folders,
  template_name = "mni_icbm152_nlin_asym_09b",
  any_mni152 = TRUE
)
}
\arguments{
\item{subject}{'RAVE' subject instance or character; see
\code{\link{as_rave_subject}}}

\item{atlas_folders}{paths to the atlas folders}

\item{template_name}{template name where the atlases are created;
see \code{\link{get_available_morph_to_template}} to get available templates
and \code{\link{cmd_run_yael_preprocess}} to generate transforms to the
templates}

\item{any_mni152}{if the template is 'MNI152', then whether to check other
template folders; default is true}
}
\value{
The function returns nothing, but will create a folder named
\code{'atlases'} under raw subject \code{'rave-imaging'} folder.
}
\description{
Reverse-transform from template atlases via volume mapping
}
\examples{

# Please check out https://rave.wiki to configure Python for RAVE
# or run ravemanager::configure_python()
\dontrun{

generate_atlases_from_template("YAEL/yael_demo_001")

}


}
