% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-constants.R
\docType{data}
\name{ravecore-constants}
\alias{ravecore-constants}
\alias{LOCATION_TYPES}
\alias{SIGNAL_TYPES}
\alias{IMPORT_FORMATS}
\title{'RAVE' constants}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{list} of length 7.
}
\usage{
LOCATION_TYPES

SIGNAL_TYPES

IMPORT_FORMATS
}
\description{
Constant variables
}
\keyword{datasets}
