% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-constants.R
\docType{data}
\name{ravecore-constants}
\alias{ravecore-constants}
\alias{LOCATION_TYPES}
\alias{SIGNAL_TYPES}
\alias{IMPORT_FORMATS}
\alias{YAEL_IMAGE_TYPES}
\alias{MNI305_to_MNI152}
\title{'RAVE' constants}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{list} of length 7.

An object of class \code{character} of length 10.

An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
LOCATION_TYPES

SIGNAL_TYPES

IMPORT_FORMATS

YAEL_IMAGE_TYPES

MNI305_to_MNI152
}
\description{
Constant variables
}
\keyword{datasets}
